This PR makes a number of changes to the local chain and net implementations as well as user-facing commands (via the CLI).

Through implementation, less is exposed (through further reducing the surface area of the API), while ensuring the local implementation is representative of our experiences with non-local work.

Lastly, a race is addressed, which occurred when groups would get registered before their pending groups would be fully updated.

+++ pkg/thresholdgroup/member.go
@@ -227,7 +227,7 @@ func (mc *memberCore) MemberID() string {
 // `dishonestThreshold` >= `groupSize` / 2, as the distributed key generation and threshold
 // signature algorithm security breaks down past that point.
 func NewMember(id string, dishonestThreshold int, groupSize int) (*LocalMember, error) {
-	if dishonestThreshold >= groupSize/2 {
+	if float64(dishonestThreshold) >= float64(groupSize)/2 {
 		return nil, fmt.Errorf(
 			"threshold %v >= %v / 2, so group security cannot be guaranteed",
 			dishonestThreshold,
